+++ app/controllers/hyrax/generic_works_controller.rb
@@ -7,6 +7,7 @@ module Hyrax
   class GenericWorksController < ApplicationController
     # Adds Hyrax behaviors to the controller.
     include Hyrax::WorksControllerBehavior
+    include Hyku::WorksControllerBehavior
     include Hyrax::BreadcrumbsForWorks
     self.curation_concern_type = ::GenericWork
