+++ app/controllers/hyrax/pages_controller_decorator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# OVERRIDE: Hyrax v5.0.0
+# OVERRIDE: Hyrax v5.0.0rc2
 # - add inject_theme_views method for theming
 # - add homepage presenter for access to feature flippers
 # - add access to content blocks in the show method
@@ -8,17 +8,22 @@
 
 module Hyrax
   # Shows the about and help page
-  class PagesController < ApplicationController
-    load_and_authorize_resource class: ContentBlock, except: :show
-    layout :pages_layout
+  module PagesControllerDecorator
+    extend ActiveSupport::Concern
 
     # OVERRIDE: Add for theming
     # Adds Hydra behaviors into the application controller
     include Blacklight::SearchContext
     include Blacklight::AccessControls::Catalog
 
-    # OVERRIDE: Adding inject theme views method for theming
-    around_action :inject_theme_views
+    prepended do
+      # OVERRIDE: Adding inject theme views method for theming
+      around_action :inject_theme_views
+
+      # OVERRIDE: Hyrax v5.0.0rc2 Add for theming
+      class_attribute :presenter_class
+      self.presenter_class = Hyrax::HomepagePresenter
+    end
 
     # OVERRIDE: Add for theming
     # The search builder for finding recent documents
@@ -27,14 +32,8 @@ module Hyrax
       Hyrax::HomepageSearchBuilder
     end
 
-    # OVERRIDE: Hyrax v3.4.0 Add for theming
-    class_attribute :presenter_class
-    self.presenter_class = Hyrax::HomepagePresenter
-
-    helper Hyrax::ContentBlockHelper
-
     def show
-      @page = ContentBlock.for(params[:key])
+      super
 
       # OVERRIDE: Additional for theming
       @presenter = presenter_class.new(current_ability, collections)
@@ -46,44 +45,7 @@ module Hyrax
       @announcement_text = ContentBlock.for(:announcement)
     end
 
-    def edit
-      add_breadcrumb t(:'hyrax.controls.home'), root_path
-      add_breadcrumb t(:'hyrax.dashboard.breadcrumbs.admin'), hyrax.dashboard_path
-      add_breadcrumb t(:'hyrax.admin.sidebar.configuration'), '#'
-      add_breadcrumb t(:'hyrax.admin.sidebar.pages'), hyrax.edit_pages_path
-    end
-
-    def update
-      respond_to do |format|
-        if @page.update(value: update_value_from_params)
-          redirect_path = "#{hyrax.edit_pages_path}##{params[:content_block].keys.first}"
-          format.html { redirect_to redirect_path, notice: t(:'hyrax.pages.updated') }
-        else
-          format.html { render :edit }
-        end
-      end
-    end
-
     private
-
-    def permitted_params
-      params.require(:content_block).permit(:about,
-                                            :agreement,
-                                            :help,
-                                            :terms)
-    end
-
-    # When a request comes to the controller, it will be for one and
-    # only one of the content blocks. Params always looks like:
-    #   {'about_page' => 'Here is an awesome about page!'}
-    # So reach into permitted params and pull out the first value.
-    def update_value_from_params
-      permitted_params.values.first
-    end
-
-    def pages_layout
-      action_name == 'show' ? 'homepage' : 'hyrax/dashboard'
-    end
 
     # OVERRIDE: return collections for theming
     # Return 6 collections, sorts by title
@@ -115,3 +77,5 @@ module Hyrax
     end
   end
 end
+
+Hyrax::PagesController.prepend(Hyrax::PagesControllerDecorator)
